# docs/swagger.yaml
basePath: /api
definitions:
  Attachment:
    properties:
      filename:
        type: string
      filetype:
        type: string
      url:
        type: string
    type: object
...
  Member:
    properties:
      color:
        type: string
      createdAt:
        type: string
      id:
        type: string
      image:
        type: string
      isFriend:
        type: boolean
      isOnline:
        type: boolean
      nickname:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  Message:
    properties:
      attachment:
        $ref: '#/definitions/Attachment'
      createdAt:
        type: string
      id:
        type: string
      text:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/Member'
    type: object
  MessageRequest:
    properties:
      file:
        format: binary
        type: string
      text:
        type: string
    type: object
...
  model.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/model.HttpError'
    type: object
  model.ErrorsResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/model.FieldError'
        type: array
    type: object
  model.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
...
  model.HttpError:
    properties:
      message:
        type: string
      type:
        type: string
    type: object
...
  model.Success:
    properties:
      success:
        type: boolean
    type: object
...
host: localhost:4000
info:
  contact:
    email: khoingoc456@gmail.com
    name: KhoiPN
  description: MurmurChatapp REST API Specs. This service uses sessions for authentication
  title: MurmurChatapp API
  version: "1.0"
paths:
...
  /messages/{channelId}:
    get:
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Cursor Pagination using the createdAt field
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Message'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get Channel Messages
      tags:
      - Messages
    post:
      parameters:
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      - description: Create Message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Create Messages
      tags:
      - Messages
  /messages/{messageId}:
    delete:
      parameters:
      - description: Message ID
        in: path
        name: messageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Delete Messages
      tags:
      - Messages
    put:
      parameters:
      - description: Message ID
        in: path
        name: messageId
        required: true
        type: string
      - description: Edit Message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Edit Messages
      tags:
      - Messages
swagger: "2.0"